Idea de hecho consensuado, falta pulir

+++ agregador/src/main/java/ar/edu/utn/frba/ddsi/agregador/models/entities/repositories/HechosRepository.java
@@ -70,7 +70,7 @@ public class HechosRepository {
                 .toList();
     }
 
-    public Double countFuentes() {
+    public Long countFuentes() {
         return fuentes.stream().count();
     }
 }
